True MP3 size is preserved.

+++ helpers/ws.php
@@ -138,7 +138,7 @@ function get_mp3s($id=null) {
             'id'           => $mp3['id'],
             'ressource_id' => $mp3['Reference'],
             'url'          => $url,
-            'size'         => (int)$mp3['Taille']
+            'size'         => (double)$mp3['Taille']
         );
     }
